requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
greenlet==1.1.2
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.2
jupyter-core==4.9.2
logging==0.4.9.6
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
